+++ src/ml_tooling/plots/__init__.py
@@ -6,6 +6,7 @@ from .prediction_error import plot_prediction_error
 from .residuals import plot_residuals
 from .roc_auc import plot_roc_auc
 from .precision_recall_curve import plot_pr_curve
+from .learning_curve import plot_learning_curve
 from .validation_curve import plot_validation_curve
 
 __all__ = [
@@ -17,5 +18,6 @@ __all__ = [
     "plot_residuals",
     "plot_roc_auc",
     "plot_pr_curve",
+    "plot_learning_curve",
     "plot_validation_curve",
 ]