+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 - Storage .save method now only takes filename as parameter
 - Handles storage loading of paths outputted from the Storage .get_list method
 - Handles case when Dataset does not have a `y` value
+- Added `plot_learning_curve` and corresponding `result.plot.learning_curve`
 - Added `plot_validation_curve` and corresponding `result.plot.validation_curve`
+
 
 # v0.9.2
   - Bug fix for logging when feature unions (DFFeatureUnion) had tuples